format code using prettier

+++ lib/ds-note/src/note.t.sol
@@ -18,16 +18,48 @@ import "../../ds-test/src/test.sol";
 import "./note.sol";
 
 contract DSNoteTest is DSTest, DSNote {
-    function setUp() public { }
-    function test_0() public { this.foo(); }
-    function test_1() public { this.foo(1); }
-    function test_2() public { this.foo(1, 2); }
-    function test_3() public { this.foo(1, 2, 3); }
-    function test_4() public { this.bar.value(uint(0x1234))(); }
+    function setUp() public {}
+
+    function test_0() public {
+        this.foo();
+    }
+
+    function test_1() public {
+        this.foo(1);
+    }
+
+    function test_2() public {
+        this.foo(1, 2);
+    }
+
+    function test_3() public {
+        this.foo(1, 2, 3);
+    }
+
+    function test_4() public {
+        this.bar.value(uint256(0x1234))();
+    }
 
     function foo() public note {}
-    function foo(uint a) public note { a; }
-    function foo(uint a, uint b) public note { a; b; }
-    function foo(uint a, uint b, uint c) public note { a; b; c; }
+
+    function foo(uint256 a) public note {
+        a;
+    }
+
+    function foo(uint256 a, uint256 b) public note {
+        a;
+        b;
+    }
+
+    function foo(
+        uint256 a,
+        uint256 b,
+        uint256 c
+    ) public note {
+        a;
+        b;
+        c;
+    }
+
     function bar() public payable note {}
 }
