+++ src/borrower/feed/test/buckets.t.sol
@@ -18,43 +18,40 @@ import "../../../../lib/galaxy-math/src/math.sol";
 
 import "./../buckets.sol";
 
-
 contract Hevm {
     function warp(uint256) public;
 }
 
 // Contract for testing the abstract Buckets contract
 contract BucketList is Buckets, Math {
-    function uniqueDayTimestamp(uint timestamp) public pure returns (uint) {
-        return (1 days) * (timestamp/(1 days));
+    function uniqueDayTimestamp(uint256 timestamp) public pure returns (uint256) {
+        return (1 days) * (timestamp / (1 days));
     }
 
-    function add(uint timestamp, uint value) public {
+    function add(uint256 timestamp, uint256 value) public {
         addBucket(uniqueDayTimestamp(timestamp), value);
     }
 
-    function remove(uint timestamp) public {
+    function remove(uint256 timestamp) public {
         removeBucket(uniqueDayTimestamp(timestamp));
     }
 
-    function calcSum() public view returns (uint) {
-        uint currDate = firstBucket;
-        uint sum = 0;
+    function calcSum() public view returns (uint256) {
+        uint256 currDate = firstBucket;
+        uint256 sum = 0;
 
         if (currDate == 0) {
             //empty list
             return 0;
         }
 
-        while(currDate != NullDate)
-        {
+        while (currDate != NullDate) {
             sum = safeAdd(sum, buckets[currDate].value);
             currDate = buckets[currDate].next;
         }
         return sum;
     }
 }
-
 
 contract BucketTest is DSTest, Math {
     Hevm hevm;
@@ -68,29 +65,29 @@ contract BucketTest is DSTest, Math {
         assertEq(buckets.firstBucket(), 0);
         assertEq(buckets.lastBucket(), 0);
 
-        uint amount = 100 ether;
+        uint256 amount = 100 ether;
 
         // add first element
         buckets.add(now + 5 days, amount);
         assertEq(buckets.calcSum(), 100 ether);
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  5 days));
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  5 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 5 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 5 days));
 
         // add second bucket after first
         buckets.add(now + 10 days, amount);
         assertEq(buckets.calcSum(), 200 ether);
         // still same first bucket
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  5 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 5 days));
         // new last bucket
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  10 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 10 days));
 
         // add before first bucket
         buckets.add(now + 3 days, amount);
         assertEq(buckets.calcSum(), 300 ether);
         // new first bucket
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  3 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 3 days));
         // same bucket
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  10 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 10 days));
     }
 
     function testAddBuckets() public {
@@ -102,9 +99,9 @@ contract BucketTest is DSTest, Math {
         // remove bucket in the middle
         buckets.remove(now + 5 days);
         assertEq(buckets.calcSum(), 200 ether);
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  3 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 3 days));
         // same bucket
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  10 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 10 days));
     }
 
     function testRemoveLastBucket() public {
@@ -112,9 +109,9 @@ contract BucketTest is DSTest, Math {
         // remove bucket in the middle
         buckets.remove(now + 10 days);
         assertEq(buckets.calcSum(), 200 ether);
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  3 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 3 days));
         // same bucket
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  5 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 5 days));
     }
 
     function testRemoveFirstBucket() public {
@@ -122,9 +119,9 @@ contract BucketTest is DSTest, Math {
         // remove bucket in the middle
         buckets.remove(now + 3 days);
         assertEq(buckets.calcSum(), 200 ether);
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  5 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 5 days));
         // same bucket
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  10 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 10 days));
     }
 
     function testRemoveAllBuckets() public {
@@ -132,8 +129,8 @@ contract BucketTest is DSTest, Math {
         buckets.remove(now + 5 days);
         buckets.remove(now + 10 days);
         assertEq(buckets.calcSum(), 100 ether);
-        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now +  3 days));
-        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now +  3 days));
+        assertEq(buckets.firstBucket(), buckets.uniqueDayTimestamp(now + 3 days));
+        assertEq(buckets.lastBucket(), buckets.uniqueDayTimestamp(now + 3 days));
 
         // remove last one
         buckets.remove(now + 3 days);
