+++ src/borrower/test/deployer.t.sol
@@ -16,17 +16,16 @@
 pragma solidity >=0.5.15 <0.6.0;
 
 import "../../../lib/ds-test/src/test.sol";
-import { Title } from "../../../lib/galaxy-title/src/title.sol";
-import { TitleFab } from "../fabs/title.sol";
-import { PileFab } from "../fabs/pile.sol";
-import { ShelfFab} from "../fabs/shelf.sol";
-import { NAVFeedFab } from "../fabs/navfeed.sol";
-import { NFTFeedFab } from "../fabs/nftfeed.sol";
-import { CollectorFab } from "../fabs/collector.sol";
-
+import {Title} from "../../../lib/galaxy-title/src/title.sol";
+import {TitleFab} from "../fabs/title.sol";
+import {PileFab} from "../fabs/pile.sol";
+import {ShelfFab} from "../fabs/shelf.sol";
+import {NAVFeedFab} from "../fabs/navfeed.sol";
+import {NFTFeedFab} from "../fabs/nftfeed.sol";
+import {CollectorFab} from "../fabs/collector.sol";
 
 import "../deployer.sol";
-import { SimpleToken } from "../../test/simple/token.sol";
+import {SimpleToken} from "../../test/simple/token.sol";
 
 contract DeployerTest is DSTest {
     Title nft;
@@ -46,11 +45,22 @@ contract DeployerTest is DSTest {
         pilefab = new PileFab();
         collectorFab = new CollectorFab();
         feedFab = new NAVFeedFab();
-   }
+    }
 
     function testBorrowerDeploy() public logs_gas {
-        uint discountRate = uint(1000000342100000000000000000);
-        BorrowerDeployer deployer = new BorrowerDeployer(address(0), address(titlefab), address(shelffab), address(pilefab), address(collectorFab), address(feedFab), address(dai), "Test", "TEST", discountRate);
+        uint256 discountRate = uint256(1000000342100000000000000000);
+        BorrowerDeployer deployer = new BorrowerDeployer(
+            address(0),
+            address(titlefab),
+            address(shelffab),
+            address(pilefab),
+            address(collectorFab),
+            address(feedFab),
+            address(dai),
+            "Test",
+            "TEST",
+            discountRate
+        );
 
         deployer.deployTitle();
         deployer.deployPile();
